+++ src/components/layout/header/StickyHeader.tsx
@@ -1,8 +1,8 @@
 import React from "react";
+import classnames from "classnames";
 
 import { Header } from "./Header.js";
 import styles from "./StickyHeader.module.scss";
-import classnames from "classnames";
 
 const err = new TypeError("StickyHeader needs two children - Header and StickyHeader.Content");