+++ src/index.ts
@@ -1,6 +1,11 @@
 export * from "./components/form/Checkbox.js";
+
+export * from "./components/icons/Icon.js";
+
 export * from "./components/layout/header/Header.js";
 export * from "./components/layout/header/StickyHeader.js";
 export * from "./components/layout/list/List.js";
 export * from "./components/layout/list/Item.js";
+
+export * from "./components/ui/button/Button.js";
 export * from "./components/ui/directionPad/Pad.js";