+++ src/types.ts
@@ -1,0 +1,7 @@
+import type React from "react";
+
+type AnyComponent = React.ComponentClass | React.FC;
+
+export type {
+    AnyComponent,
+};